refactored structure of controllers and repositories
