+++ src/main/java/com/dvlcube/app/aspect/RestAspect.java
@@ -8,7 +8,7 @@ import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.springframework.context.annotation.Configuration;
 
-import com.dvlcube.app.rest.StatService;
+import com.dvlcube.app.controller.StatController;
 import com.dvlcube.utils.aspects.stats.StatsAspect;
 import com.dvlcube.utils.interfaces.MxAspect;
 import com.dvlcube.utils.interfaces.MxBean;
@@ -17,7 +17,7 @@ import com.dvlcube.utils.interfaces.MxBean;
  * Logs all HTTP requests, generating stats about the time it takes to run every
  * method.
  * 
- * @see StatService
+ * @see StatController
  * @since 13 de fev de 2019
  * @author Ulisses Lima
  */
