+++ src/main/java/com/dvlcube/app/repository/SkillRepository.java
@@ -1,10 +1,10 @@
-package com.dvlcube.app.jpa.repo;
+package com.dvlcube.app.repository;
 
 import org.springframework.stereotype.Repository;
 
 import com.dvlcube.app.manager.data.SkillBean;
-import com.dvlcube.app.jpa.BasicRepository;
-import com.dvlcube.app.jpa.DvlRepository;
+import com.dvlcube.app.repository.BasicRepository;
+import com.dvlcube.app.repository.DvlRepository;
 
 /**
  * @since 4 de jun de 2019
